pyproject.toml
[tool.poetry]
name = "DjangoStartKit"
version = "v0.01"
description = "Django Start Kit"
authors = ["phoenix <hypofiasco@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.12"
mysqlclient = "2.0.3"
python-dotenv = "0.12.0"
redis = "3.5.3"
happybase = "1.2.0"
ujson = "2.0"
httpx = "0.18.1"
requests = "2.25.1"
pycryptodomex = "3.10.1"
pyopenssl = "22.0.0"
passlib = "1.7"
email-validator = "^1.0.5"
django = "3.2.9"
djangorestframework = "3.12.4"
django-filter = "21.1"
drf-yasg = "1.20.0"
django-cors-headers = "3.10.0"
pydantic = "1.8.2"
djangorestframework-jwt = "1.11.0"
django-simple-captcha = "0.5.14"
typer = "0.3.2"
pypinyin = "0.40.0"
redisearch = "2.1.1"
cachetools = "4.2.4"
kubernetes = "9.0.0"
gevent = "21.1.2"
gunicorn = "20.1.0"
pyyaml = "6.0"
channels = "3.0.4"
channels-redis = "3.4.1"
asgiref = "3.3.4"
uvicorn = "0.18.2"
websockets = "10.3"

[tool.poetry.dev-dependencies]
# poetry install --no-dev
factory-boy = "3.2.1"
black = "19.10b0"
flake8 = "3.9.1"
isort = "5.1.4"
pytest = "6.2.4"

[[tool.poetry.source]]
name = "aliyun"
default = true
url = "https://mirrors.aliyun.com/pypi/simple/"

# If poetry install failed, excute follow shell command
# shell
# >>> poetry cache clear . --all
# >>> rm poetry.lock
# >>> poetry install -vvv
